get_job_info prep

+++ src/OSmOSE/core/job.py
@@ -3,6 +3,7 @@ import glob
 import os
 import json
 import time
+from dateutil.parser import parse
 import tomlkit
 import subprocess
 from uuid import uuid4
@@ -265,19 +266,19 @@ class Job_builder:
         today = datetime.today().strftime("%Y_%m_%d")
 
         if logdir is None:
-            logdir = pwd.joinpath("log_job",today)
-            jobdir = pwd.joinpath("ongoing_jobs",today)
-            make_path(logdir, mode=DPDEFAULT)
-            make_path(jobdir, mode=DPDEFAULT)
+            self.logdir = pwd.joinpath("log_job",today)
+            self.jobdir = pwd.joinpath("ongoing_jobs",today)
+            make_path(self.logdir, mode=DPDEFAULT)
+            make_path(self.jobdir, mode=DPDEFAULT)
         else:
-            logdir = logdir.joinpath(today)
-            make_path(logdir, mode=DPDEFAULT)
-            jobdir = logdir
+            self.logdir = logdir.joinpath(today)
+            make_path(self.logdir, mode=DPDEFAULT)
+            self.jobdir = self.logdir
         
 
         job_file = ["#!/bin/bash"]
 
-        date_id = datetime.now().strftime('%H-%M-%S')
+        date_id = datetime.now.strftime('%H-%M-%S')
         # region Build header
         #! HEADER
         # Getting all header variables from the parameters or the configuration defaults
@@ -291,10 +292,10 @@ class Job_builder:
         if not outfile:
             outfile = self.outfile
 
-        outfile = logdir.joinpath(outfile)
+        outfile = self.logdir.joinpath(outfile)
         if not errfile:
             errfile = self.errfile
-        errfile = logdir.joinpath(errfile)
+        errfile = self.logdir.joinpath(errfile)
 
         if not queue:
             queue = self.queue if not preset else job_preset.queue
@@ -380,9 +381,9 @@ class Job_builder:
         job_file.append(f"python {script_path} {script_args}")
 
         #! FOOTER
-        outfilename = f"{jobname}_{date_id}_{job_scheduler}_{len(os.listdir(jobdir))}.pbs"
+        outfilename = f"{jobname}_{date_id}_{job_scheduler}_{len(os.listdir(self.jobdir))}.pbs"
 
-        job_file_path = jobdir.joinpath(outfilename)
+        job_file_path = self.jobdir.joinpath(outfilename)
 
         # job_file.append(f"\nchmod 444 {outfile} {errfile}")
         job_file.append(f"\nrm {job_file_path}\n")
@@ -462,6 +463,7 @@ class Job_builder:
 
             if jobinfo in self.prepared_jobs:
                 self.__prepared_jobs.remove(jobinfo)
+            jobinfo["id"] = jobid
             self.__ongoing_jobs.append(jobinfo)
 
         return jobid_list
@@ -518,6 +520,28 @@ class Job_builder:
 
         print(res)
 
+    def search_jobs(self, *, date:str, log_dir: Path= None):
+        top_dir = log_dir if log_dir is not None else self.logdir
+        if top_dir is None:
+            raise ValueError("If no job has been submitted by this instance of Job_builder, then log_dir must be specified.")
+        
+        date = parse(date, fuzzy=True)
+        foldername = datetime.stftime(date, "%Y_%m_%d")
+
+        return [output for output in os.listdir(top_dir.joinpath(foldername)) if output.endswith(".out")]
+
+    def get_job_info(self, *, job_identifier:str):
+        jobinfo = self.get_ongoing_job(job_identifier=job_identifier)
+        
+        if "Torque" in str(jobinfo["path"]):
+            info = subprocess.run(["qstat", "-fx", jobinfo["id"]], stdout=subprocess.PIPE).stdout.decode("utf-8")
+            elapsed = info.split("resources_used.walltime = ")[1].split("\n")[0]
+        elif "Slurm" in str(jobinfo["path"]):
+            subprocess.run(["squeue", jobinfo["id"]])
+
+        
+
+
     def read_output_file(
         self, *, job_name: str = None, outtype: Literal["out", "err"] = "out", job_file_name: str = None
     ):
@@ -563,12 +587,31 @@ class Job_builder:
                 If set to True, will clear the job history of the job builder."""
         if all_jobs:
             for jobinfo in self.all_jobs:
-                jobinfo["path"].unlink()
+                if jobinfo["path"].exists():
+                    jobinfo["path"].unlink()
 
         for jobinfo in self.ongoing_jobs:
             res = subprocess.run(f"qstat {jobinfo['job_id']}", shell=True, stdout=subprocess.PIPE).stdout.decode("utf-8").rstrip("\n")
-            if f"{jobinfo['job_id']} has finished" in res:
+            if f"{jobinfo['job_id']} has finished" in res and jobinfo["path"].exists():
                 jobinfo["path"].unlink()
+                
+    def delete_job(self, job_identifier:str):
+        jobinfo = self.get_ongoing_job(job_identifier=job_identifier)
+        
+        if "Torque" in str(jobinfo["path"]):
+            subprocess.run(["qdel", jobinfo["id"]])
+        elif "Slurm" in str(jobinfo["path"]):
+            subprocess.run(["scancel", jobinfo["id"]])
+        self.__ongoing_jobs.remove(jobinfo)
+        self.__cancelled_jobs.append(jobinfo)
+
+    def get_ongoing_job(self, job_identifier:str):
+        try:
+            job_id = get_dict_index_in_list(self.ongoing_jobs, "job_name", job_identifier.rstrip())
+        except ValueError:
+            job_id = get_dict_index_in_list(self.ongoing_jobs, "id", job_identifier.rstrip())
+
+        return self.ongoing_jobs[job_id]
 
 def get_dict_index_in_list(list: list, key: str, value: any) -> int:
     for i, d in enumerate(list):
